+++ pytdx/reader/__init__.py
@@ -3,6 +3,7 @@ from pytdx.reader.min_bar_reader import TdxMinBarReader
 from pytdx.reader.lc_min_bar_reader import TdxLCMinBarReader
 from pytdx.reader.exhq_daily_bar_reader import TdxExHqDailyBarReader
 from pytdx.reader.gbbq_reader import GbbqReader
+from pytdx.reader.block_reader import BlockReader
 
 __all__ = [
     'TdxDailyBarReader',
@@ -12,4 +13,5 @@ __all__ = [
     'TdxLCMinBarReader',
     'TdxExHqDailyBarReader',
     'GbbqReader',
+    'BlockReader',
 ]